Track_Database/requirements.txt
alembic~=1.6.2
Faker==8.2.0
psycopg2-binary~=2.8.6
python-dotenv~=0.17.1
SQLAlchemy~=1.4.15